+++ src/main/java/com/intrinio/realtime/QuoddTradeQuote.java
@@ -2,74 +2,72 @@ package com.intrinio.realtime;
 
 import org.json.JSONObject;
 
-import java.math.BigDecimal;
-
 public class QuoddTradeQuote implements Quote {
-    private BigDecimal changePrice4d = null;
-    private BigDecimal dayHigh4d = null;
-    private BigDecimal dayHighTime = null;
-    private BigDecimal dayLow4d = null;
-    private BigDecimal dayLowTime = null;
-    private BigDecimal extChangePrice4d = null;
-    private BigDecimal extLastPrice4d = null;
-    private BigDecimal extPercentChange4d = null;
+    private Double changePrice4d = null;
+    private Double dayHigh4d = null;
+    private Double dayHighTime = null;
+    private Double dayLow4d = null;
+    private Double dayLowTime = null;
+    private Double extChangePrice4d = null;
+    private Double extLastPrice4d = null;
+    private Double extPercentChange4d = null;
     private String extTradeExchange = null;
-    private BigDecimal extTradeTime = null;
-    private BigDecimal extTradeVolume = null;
+    private Double extTradeTime = null;
+    private Double extTradeVolume = null;
     private String extUpDown = null;
     private Boolean isHalted = null;
     private Boolean isShortRestricted = null;
-    private BigDecimal lastPrice4d = null;
-    private BigDecimal openPrice4d = null;
-    private BigDecimal openTime = null;
-    private BigDecimal openVolume = null;
-    private BigDecimal percentChange4d = null;
-    private BigDecimal prevClose4d = null;
-    private BigDecimal protocolId = null;
+    private Double lastPrice4d = null;
+    private Double openPrice4d = null;
+    private Double openTime = null;
+    private Double openVolume = null;
+    private Double percentChange4d = null;
+    private Double prevClose4d = null;
+    private Double protocolId = null;
     private String rootTicker = null;
-    private BigDecimal rtl = null;
+    private Double rtl = null;
     private String ticker = null;
-    private BigDecimal totalVolume = null;
+    private Double totalVolume = null;
     private String tradeExchange = null;
-    private BigDecimal tradeTime = null;
-    private BigDecimal tradeVolume = null;
+    private Double tradeTime = null;
+    private Double tradeVolume = null;
     private String upDown = null;
-    private BigDecimal volumePlus = null;
-    private BigDecimal vwap4d = null;
+    private Double volumePlus = null;
+    private Double vwap4d = null;
 
     public QuoddTradeQuote(JSONObject message) {
         if (message.has("change_price_4d")) {
-            this.changePrice4d = message.getBigDecimal("change_price_4d");
+            this.changePrice4d = message.getDouble("change_price_4d");
         }
         if (message.has("day_high_4d")) {
-            this.dayHigh4d = message.getBigDecimal("day_high_4d");
+            this.dayHigh4d = message.getDouble("day_high_4d");
         }
         if (message.has("day_high_time")) {
-            this.dayHighTime = message.getBigDecimal("day_high_time");
+            this.dayHighTime = message.getDouble("day_high_time");
         }
         if (message.has("day_low_4d")) {
-            this.dayLow4d = message.getBigDecimal("day_low_4d");
+            this.dayLow4d = message.getDouble("day_low_4d");
         }
         if (message.has("day_low_time")) {
-            this.dayLowTime = message.getBigDecimal("day_low_time");
+            this.dayLowTime = message.getDouble("day_low_time");
         }
         if (message.has("ext_change_price_4d")) {
-            this.extChangePrice4d = message.getBigDecimal("ext_change_price_4d");
+            this.extChangePrice4d = message.getDouble("ext_change_price_4d");
         }
         if (message.has("ext_last_price_4d")) {
-            this.extLastPrice4d = message.getBigDecimal("ext_last_price_4d");
+            this.extLastPrice4d = message.getDouble("ext_last_price_4d");
         }
         if (message.has("ext_percent_change_4d")) {
-            this.extPercentChange4d = message.getBigDecimal("ext_percent_change_4d");
+            this.extPercentChange4d = message.getDouble("ext_percent_change_4d");
         }
         if (message.has("ext_trade_exchange")) {
             this.extTradeExchange = message.getString("ext_trade_exchange");
         }
         if (message.has("ext_trade_time")) {
-            this.extTradeTime = message.getBigDecimal("ext_trade_time");
+            this.extTradeTime = message.getDouble("ext_trade_time");
         }
         if (message.has("ext_trade_volume")) {
-            this.extTradeVolume = message.getBigDecimal("ext_trade_volume");
+            this.extTradeVolume = message.getDouble("ext_trade_volume");
         }
         if (message.has("ext_up_down")) {
             this.extUpDown = message.getString("ext_up_down");
@@ -81,87 +79,87 @@ public class QuoddTradeQuote implements Quote {
             this.isShortRestricted = message.getBoolean("is_short_restricted");
         }
         if (message.has("last_price_4d")) {
-            this.lastPrice4d = message.getBigDecimal("last_price_4d");
+            this.lastPrice4d = message.getDouble("last_price_4d");
         }
         if (message.has("open_price_4d")) {
-            this.openPrice4d = message.getBigDecimal("open_price_4d");
+            this.openPrice4d = message.getDouble("open_price_4d");
         }
         if (message.has("open_time")) {
-            this.openTime = message.getBigDecimal("open_time");
+            this.openTime = message.getDouble("open_time");
         }
         if (message.has("open_volume")) {
-            this.openVolume = message.getBigDecimal("open_volume");
+            this.openVolume = message.getDouble("open_volume");
         }
         if (message.has("percent_change_4d")) {
-            this.percentChange4d = message.getBigDecimal("percent_change_4d");
+            this.percentChange4d = message.getDouble("percent_change_4d");
         }
         if (message.has("prev_close_4d")) {
-            this.prevClose4d = message.getBigDecimal("prev_close_4d");
+            this.prevClose4d = message.getDouble("prev_close_4d");
         }
         if (message.has("protocol_id")) {
-            this.protocolId = message.getBigDecimal("protocol_id");
+            this.protocolId = message.getDouble("protocol_id");
         }
         if (message.has("root_ticker")) {
             this.rootTicker = message.getString("root_ticker");
         }
         if (message.has("rtl")) {
-            this.rtl = message.getBigDecimal("rtl");
+            this.rtl = message.getDouble("rtl");
         }
         if (message.has("ticker")) {
             this.ticker = message.getString("ticker");
         }
         if (message.has("total_volume")) {
-            this.totalVolume = message.getBigDecimal("total_volume");
+            this.totalVolume = message.getDouble("total_volume");
         }
         if (message.has("trade_exchange")) {
             this.tradeExchange = message.getString("trade_exchange");
         }
         if (message.has("trade_time")) {
-            this.tradeTime = message.getBigDecimal("trade_time");
+            this.tradeTime = message.getDouble("trade_time");
         }
         if (message.has("trade_volume")) {
-            this.tradeVolume = message.getBigDecimal("trade_volume");
+            this.tradeVolume = message.getDouble("trade_volume");
         }
         if (message.has("up_down")) {
             this.upDown = message.getString("up_down");
         }
         if (message.has("volume_plus")) {
-            this.volumePlus = message.getBigDecimal("volume_plus");
+            this.volumePlus = message.getDouble("volume_plus");
         }
         if (message.has("vwap_4d")) {
-            this.vwap4d = message.getBigDecimal("vwap_4d");
+            this.vwap4d = message.getDouble("vwap_4d");
         }
     }
 
-    public BigDecimal getChangePrice4d() {
+    public Double getChangePrice4d() {
         return changePrice4d;
     }
 
-    public BigDecimal getDayHigh4d() {
+    public Double getDayHigh4d() {
         return dayHigh4d;
     }
 
-    public BigDecimal getDayHighTime() {
+    public Double getDayHighTime() {
         return dayHighTime;
     }
 
-    public BigDecimal getDayLow4d() {
+    public Double getDayLow4d() {
         return dayLow4d;
     }
 
-    public BigDecimal getDayLowTime() {
+    public Double getDayLowTime() {
         return dayLowTime;
     }
 
-    public BigDecimal getExtChangePrice4d() {
+    public Double getExtChangePrice4d() {
         return extChangePrice4d;
     }
 
-    public BigDecimal getExtLastPrice4d() {
+    public Double getExtLastPrice4d() {
         return extLastPrice4d;
     }
 
-    public BigDecimal getExtPercentChange4d() {
+    public Double getExtPercentChange4d() {
         return extPercentChange4d;
     }
 
@@ -169,11 +167,11 @@ public class QuoddTradeQuote implements Quote {
         return extTradeExchange;
     }
 
-    public BigDecimal getExtTradeTime() {
+    public Double getExtTradeTime() {
         return extTradeTime;
     }
 
-    public BigDecimal getExtTradeVolume() {
+    public Double getExtTradeVolume() {
         return extTradeVolume;
     }
 
@@ -189,31 +187,31 @@ public class QuoddTradeQuote implements Quote {
         return isShortRestricted;
     }
 
-    public BigDecimal getLastPrice4d() {
+    public Double getLastPrice4d() {
         return lastPrice4d;
     }
 
-    public BigDecimal getOpenPrice4d() {
+    public Double getOpenPrice4d() {
         return openPrice4d;
     }
 
-    public BigDecimal getOpenTime() {
+    public Double getOpenTime() {
         return openTime;
     }
 
-    public BigDecimal getOpenVolume() {
+    public Double getOpenVolume() {
         return openVolume;
     }
 
-    public BigDecimal getPercentChange4d() {
+    public Double getPercentChange4d() {
         return percentChange4d;
     }
 
-    public BigDecimal getPrevClose4d() {
+    public Double getPrevClose4d() {
         return prevClose4d;
     }
 
-    public BigDecimal getProtocolId() {
+    public Double getProtocolId() {
         return protocolId;
     }
 
@@ -221,7 +219,7 @@ public class QuoddTradeQuote implements Quote {
         return rootTicker;
     }
 
-    public BigDecimal getRtl() {
+    public Double getRtl() {
         return rtl;
     }
 
@@ -229,7 +227,7 @@ public class QuoddTradeQuote implements Quote {
         return ticker;
     }
 
-    public BigDecimal getTotalVolume() {
+    public Double getTotalVolume() {
         return totalVolume;
     }
 
@@ -237,11 +235,11 @@ public class QuoddTradeQuote implements Quote {
         return tradeExchange;
     }
 
-    public BigDecimal getTradeTime() {
+    public Double getTradeTime() {
         return tradeTime;
     }
 
-    public BigDecimal getTradeVolume() {
+    public Double getTradeVolume() {
         return tradeVolume;
     }
 
@@ -249,11 +247,11 @@ public class QuoddTradeQuote implements Quote {
         return upDown;
     }
 
-    public BigDecimal getVolumePlus() {
+    public Double getVolumePlus() {
         return volumePlus;
     }
 
-    public BigDecimal getVwap4d() {
+    public Double getVwap4d() {
         return vwap4d;
     }
